+++ python/GafferSceneUI/ShaderUI.py
@@ -429,7 +429,7 @@ GafferUI.GraphEditor.plugContextMenuSignal().connect( __graphEditorPlugContextMe
 # ShaderParameterDialog
 ##########################################################################
 
-class _DuplicateIconColumn ( GafferUI.PathColumn ) :
+class _DuplicateIconColumn( GafferUI.PathColumn ) :
 
 	def __init__( self, title, property ) :
 
@@ -457,11 +457,11 @@ class _DuplicateIconColumn ( GafferUI.PathColumn ) :
 
 		return data
 
-	def headerData( self, canceller = None ) :
+	def headerData( self, rootPath, canceller = None ) :
 
 		return self.CellData( self.__title )
 
-class _ShaderInputColumn ( GafferUI.PathColumn ) :
+class _ShaderInputColumn( GafferUI.PathColumn ) :
 
 	def __init__( self, title ) :
 
@@ -492,7 +492,7 @@ class _ShaderInputColumn ( GafferUI.PathColumn ) :
 
 		return data
 
-	def headerData( self, canceller = None ) :
+	def headerData( self, rootPath, canceller = None ) :
 
 		return self.CellData( self.__title )